+++ TeamCode/src/main/java/tele/TerraOp.java
@@ -1,4 +1,4 @@
-package teleop;
+package tele;
 
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
